--- src/components/HomeReady.jsx
@@ -1,12 +1,0 @@
-
-const HomeReady = () => {
-    return (
-        <section className='bg-green-800 p-10 mt-0 md:-mt-20'>
-            <p className='text-white mt-5 text-3xl font-bold text-center'>Ready to List Your Property?</p>
-            <p className='text-gray-300 mb-10 mt-10 font-bold text-center'>Join thousands of landlords who trust RentEase to find quality tenants.</p>
-            <button className='bg-white mb-5 text-green-800 font-bold py-2 px-4 rounded mt-5 block mx-auto hover:bg-gray-200'>List Your Property</button>
-        </section>
-    )
-}
-
-export default HomeReady
