+++ src/components/LandingHero.jsx
@@ -32,11 +32,13 @@ const LandingHero = ({ onSearch }) => {
 
         {/* Search Bar */}
         <div className="mt-10 max-w-2xl mx-auto w-full">
-          <form className="flex w-full items-center rounded-full bg-white/90 shadow-2xl backdrop-blur-sm p-2">
-            <div className="md:flex hidden flex-shrink-0 pl-4 pr-2">
-              <span className="text-gray-500">
-                <FaSearchLocation />
-              </span>
+          <form
+            onSubmit={handleSubmit}
+            className="flex w-full items-center rounded-full bg-white/90 shadow-2xl backdrop-blur-sm p-2"
+          >
+            {/* Icon */}
+            <div className="flex-shrink-0 pl-4 pr-2 text-gray-500">
+              <FaSearchLocation />
             </div>
 
             {/* Input */}
@@ -50,15 +52,10 @@ const LandingHero = ({ onSearch }) => {
 
             {/* Button */}
             <button
-              className="md:flex justify-center items-center hidden flex-shrink-0 rounded-full h-12 px-8 bg-green-900 text-white text-base font-bold tracking-wide hover:bg-green-800 transition-all duration-300 cursor-pointer"
+              className="flex-shrink-0 rounded-full h-12 px-8 bg-blue-600 text-white text-base font-bold tracking-wide hover:bg-[var(--primary-color-hover)] transition-all duration-300"
               type="submit"
             >
               Search
-            </button>
-            <button type="submit" className="flex md:hidden flex-shrink-0 pl-4 pr-2">
-              <span className="text-gray-500">
-                <FaSearchLocation />
-              </span>
             </button>
           </form>
         </div>
